Distribute test cases
